pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "arc"
version = "1.0.0"
description = "ARC: Multi-Agent RL Environment for Mixed Urban Mobility"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openenv-core>=0.1.1",
    "fastapi>=0.110.0",
    "uvicorn>=0.29.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
train = [
    "trl>=0.8.0",
    "transformers>=4.40.0",
    "unsloth",
    "torch>=2.2.0",
    "wandb",
]
dev = [
    "pytest",
    "httpx",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["arc_india*"]

[project.scripts]
arc-india-train = "arc_india.train_grpo:main"
